+++ app/src/main/java/com/zarinraim/dogsearch/di/AppModule.kt
@@ -2,14 +2,14 @@ package com.zarinraim.dogsearch.di
 
 import com.zarinraim.dogsearch.data.repository.BreedRepositoryImpl
 import com.zarinraim.dogsearch.domain.repository.BreedsRepository
-import com.zarinraim.dogsearch.presentation.main.all_breeds.DogBreedsListModel
-import com.zarinraim.dogsearch.presentation.main.dog_image.BreedImageModel
+import com.zarinraim.dogsearch.feature.list.BreedsListModel
+import com.zarinraim.dogsearch.feature.detail.BreedImageModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val appModule = module {
     single<BreedsRepository> { BreedRepositoryImpl() }
-    viewModel { DogBreedsListModel(get()) }
+    viewModel { BreedsListModel(get()) }
     viewModel { (breedName: String, subBreedName: String) ->
         BreedImageModel(
             get(),
