+++ app/src/main/java/com/zarinraim/dogsearch/feature/detail/BreedImageScreen.kt
@@ -1,4 +1,4 @@
-package com.zarinraim.dogsearch.presentation.main.dog_image
+package com.zarinraim.dogsearch.feature.detail
 
 import android.content.res.Configuration
 import androidx.compose.foundation.Image
