+++ app/src/main/java/com/zarinraim/dogsearch/feature/list/BreedsListModel.kt
@@ -1,4 +1,4 @@
-package com.zarinraim.dogsearch.presentation.main.all_breeds
+package com.zarinraim.dogsearch.feature.list
 
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
@@ -8,7 +8,7 @@ import com.zarinraim.dogsearch.data.model.toBreeds
 import com.zarinraim.dogsearch.domain.repository.BreedsRepository
 import kotlinx.coroutines.launch
 
-class DogBreedsListModel(private val repo: BreedsRepository) : ViewModel() {
+class BreedsListModel(private val repo: BreedsRepository) : ViewModel() {
 
     private val _state = mutableStateOf(BreedsListState())
     val state: State<BreedsListState> = _state
