+++ app/src/main/java/com/zarinraim/dogsearch/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.zarinraim.dogsearch.presentation.ui.theme
+package com.zarinraim.dogsearch.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
