+++ app/src/main/java/com/zarinraim/dogsearch/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.zarinraim.dogsearch.presentation.ui.theme
+package com.zarinraim.dogsearch.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
